Add Cluster manager
This commit adds a class that can be used to communicate with the cluster where the containers that build the sites live.
This class has a general API that is similar to ECS's.
Right now it primarily pipes calls to ECS's API and returns a promise for the results.

The purpose of this class is to make it simple to swap in a cluster manager that runs builds on Diego when it is time to make that change.