Add CloudFoundryAPIClient
This commit adds the CloudFoundryAPIClient which introduces 2 functions for interacting with the Cloud Foundry API.

The first of these functions, `fetchBuildContainers` looks in the current space for any apps that could be used for running builds.
The second, `updateBuildContainer` updates a build containers environment and restages the container.

The CloudFoundryAPIClient is backed up by the CloudFoundryAuthClient which was introduced in a previous commit.