Check that container has build before checking buildID
There were some errors on staging resulting from the Cluster Class's `_findBuildContainer(buildID)` method.

The error message was `Cannot read property buildID of undefined` at this line:

```javascript
return container.build.buildID === buildID
```

The build is expected to be undefined when a container does not have a build running.
This commit adjusts that code so that first we check for a build to be present before comparing buildID's.