Only send build timeout message when builds timeout
Prior to this commit, the `BuildTimeoutReporter` was invoked in `Cluster.stopBuild`. This mean a timeout was reported any time a build was stopped, regardless of whether or not the build was stopped because it timed out or because it completed successfully.

This commit adds a new function for stopping builds that timeout and reporting a timeout. It also adds better test coverage for timeout behavior to prevent this error from appearing in the future.